+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [SimpleTool](https://github.com/nchekwa/simpletool-python/tree/master).
 
+## [0.0.4] - 2025-01-02
+
+### Fixed
+- select_random_api_key where value is empty string cause error
+
+### Added
+
+- add proper validation of run/execute output type (validate_tool_output decorator)
+this will allow only to use as output: Union[List[ContentT], ErrorData]
+where: ContentT = TypeVar('ContentT', ImageContent, TextContent, EmbeddedResource)
 
 ## [0.0.3] - 2025-01-02